+++ Toro.Domain/Entities/Stock.cs
@@ -8,7 +8,7 @@ namespace Toro.Domain.Entities
     public class Stock : BaseEntity
     {
         public string Symbol { get; set; }
-        public decimal CurrentPrice { get; set; }
+        public double CurrentPrice { get; set; }
 
     }
 }
